feat: Allow the tile to use for landfill to be configured
Part of my request in https://github.com/GlassBricks/StagedBlueprintPlanning/issues/12 .

Since this is a global setting there isn't a way to use different landfill tiles for different blueprints. That might be a nice thing to have but it is probably not worth cluttering up the interface for.

This does require one to lookup the name of the tile via the debug UI which isn't ideal but I don't have a better idea at this time.